+++ UIPathValidator/Validation/WorkflowValidator.cs
@@ -32,7 +32,7 @@ namespace UIPathValidator.Validation
             AddResults(new VariableNameReferee().Validate(Workflow));
             GetAndValidateInvokes();
             AddResults(new EmptyIfReferee().Validate(Workflow));
-            ValidateFlowchartActivities();
+            AddResults(new FlowchartReferee().Validate(Workflow));
             AddResults(new EmptySequenceReferee().Validate(Workflow));
             AddResults(new EmptyWhileReferee().Validate(Workflow));
             AddResults(new EmptyDoWhileReferee().Validate(Workflow));
@@ -160,92 +160,6 @@ namespace UIPathValidator.Validation
                     AddResult(new InvokeValidationResult(this.Workflow, workflow.FilePath, displayName, ValidationResultType.Error, message));
                 }
             }
-        }
-
-        private void ValidateFlowchartActivities()
-        {
-            var reader = Workflow.GetXamlReader();
-            var flowchartTags = reader.Document.Descendants(XName.Get("Flowchart", reader.Namespaces.DefaultNamespace));
-
-            foreach (var flowchartTag in flowchartTags)
-            {
-                if (flowchartTag.IsInsideCommentOut(reader.Namespaces))
-                    continue;
-
-                var startNode = flowchartTag.Element(XName.Get("Flowchart.StartNode", reader.Namespaces.DefaultNamespace));
-                var name = flowchartTag.Attribute("DisplayName")?.Value ?? "Flowchart";
-
-                if (startNode == null)
-                {
-                    var message = "Flowchart activity doens't have a Start Node.";
-                    AddResult(new FlowchartValidationResult(this.Workflow, name, ValidationResultType.Warning, message));
-                }
-                else
-                {
-                    var orphanNodes = flowchartTag.Elements(XName.Get("FlowStep", reader.Namespaces.DefaultNamespace));
-
-                    if (orphanNodes.Count() > 0)
-                    {
-                        var message = string.Format("Flowchart contains {0} node{1} that will never be reached. Either use {2} ot delete {2}.",
-                            orphanNodes.Count(),
-                            orphanNodes.Count() > 1 ? "s" : string.Empty,
-                            orphanNodes.Count() > 1 ? "them" : "it");
-                        AddResult(new FlowchartValidationResult(this.Workflow, name, ValidationResultType.Warning, message));
-                    }
-
-                    var decisionTags = startNode.Descendants(XName.Get("FlowDecision", reader.Namespaces.DefaultNamespace));
-                    var switchTags = startNode.Descendants(XName.Get("FlowSwitch", reader.Namespaces.DefaultNamespace));
-                    bool hasAnyDecisionTag = false;
-
-                    if (decisionTags.Count() > 0)
-                    {
-                        foreach (var tag in decisionTags)
-                        {
-                            if (IsDirectlyInFlowchart(tag, startNode, reader.Namespaces))
-                            {
-                                hasAnyDecisionTag = true;
-                                break;
-                            }
-                        }
-                    }
-
-                    if (!hasAnyDecisionTag && switchTags.Count() > 0)
-                    {
-                        foreach (var tag in switchTags)
-                        {
-                            if (IsDirectlyInFlowchart(tag, startNode, reader.Namespaces))
-                            {
-                                hasAnyDecisionTag = true;
-                                break;
-                            }
-                        }
-                    }
-
-                    if (!hasAnyDecisionTag)
-                    {
-                        var message = "Flowchart activity doens't have any decisions. Shouldn't this be a sequence?";
-                        AddResult(new FlowchartValidationResult(this.Workflow, name, ValidationResultType.Info, message));
-                    }
-                }
-            }
-        }
-
-        private bool IsDirectlyInFlowchart(XElement node, XElement flowchart, XmlNamespaceManager namespaces)
-        {
-            var startNodes = node.Ancestors(XName.Get("Flowchart.StartNode", namespaces.DefaultNamespace));
-            
-            if (startNodes.Count() == 0)
-                return false;
-            
-            if (!namespaces.HasNamespace("sap2010"))
-                return false;
-            
-            string firstParentNodeRefId = startNodes.First().Parent.Attribute(XName.Get("WorkflowViewState.IdRef", namespaces.LookupNamespace("sap2010")))?.Value;
-            string flowchartRefId = flowchart.Parent.Attribute(XName.Get("WorkflowViewState.IdRef", namespaces.LookupNamespace("sap2010")))?.Value;
-            if (flowchartRefId.Equals(firstParentNodeRefId))
-                return true;
-            
-            return false;
         }
     }
 }
